Change `:input-stream` to `:input` in request map
The request and response map keys should represent the role of the
data and not the format of the data. `:input` is expected to act like
`:body` in ring, it starts off as an input stream but the included
middleware will transform it into an edn representation maintaining
the same key name.